+++ lib/jxl/dec_cache.h
@@ -90,6 +90,9 @@ struct PassesDecoderState {
   // If true, rgb_output or callback output is RGBA using 4 instead of 3 bytes
   // per pixel.
   bool rgb_output_is_rgba;
+  // If true, the RGBA output will be unpremultiplied before writing to the
+  // output callback (the output buffer case is handled in ConvertToExternal).
+  bool unpremul_alpha;
 
   // Callback for line-by-line output.
   PixelCallback pixel_callback;
@@ -135,6 +138,7 @@ struct PassesDecoderState {
 
     rgb_output = nullptr;
     rgb_output_is_rgba = false;
+    unpremul_alpha = false;
     fast_xyb_srgb8_conversion = false;
     pixel_callback = PixelCallback();
     used_acs = 0;